+++ configs.js
@@ -1,5 +1,5 @@
 const configs = {
-    server_url: 'https://us-central1-app-neptunechain.cloudfunctions.net/app'
+    server_url: 'https://us-central1-app-neptunechain.cloudfunctions.net/app',
 }
 
 export default configs;
