+++ package.json
@@ -25,7 +25,7 @@
     "chart.js": "^4.4.1",
     "crypto-js": "^4.1.1",
     "ethers": "^6.8.1",
-    "firebase": "^10.4.0",
+    "firebase": "^10.8.1",
     "firebaseui": "^6.1.0",
     "font-awesome": "^4.7.0",
     "framer": "^2.3.0",
@@ -39,6 +39,7 @@
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
+    "react-firebaseui": "^6.0.0",
     "react-icons": "^4.12.0",
     "react-loader-spinner": "^5.4.5",
     "react-lottie": "^1.2.3",
