--- src/apis/contracts/index.js
@@ -1,7 +1,0 @@
-import * as MarketplaceContract from './marketplace'
-import * as VerificationContract from './verification'
-
-export {
-    VerificationContract,
-    MarketplaceContract
-}
