+++ src/smart_contracts/abis/index.js
@@ -1,0 +1,9 @@
+import MarketplaceContract  from "./marketplace_contract";
+import VerificationContract from "./verification_contract";
+import * as ContractStandards from "./standards"
+
+export {
+    MarketplaceContract,
+    VerificationContract,
+    ContractStandards
+}
