+++ src/smart_contracts/interactions/index.js
@@ -1,0 +1,7 @@
+import getVerificationInteractions from "./verification";
+import getMarketInteractions from "./marketplace";
+
+export {
+    getVerificationInteractions,
+    getMarketInteractions
+}
